refactor(view): view 관련 함수 폴더 이동 및 경로 수정
