+++ src/view/functions/printRunRacingCar.js
@@ -1,5 +1,7 @@
 import { MissionUtils } from '@woowacourse/mission-utils';
 
+import { OUTPUT } from '../../constants/constants.js';
+
 export const printRunRacingCar = async () => {
-  MissionUtils.Console.print('\n실행 결과');
+  MissionUtils.Console.print(OUTPUT.PROGRESS_RESULT);
 };
