+++ src/view/functions/progressRacingCar.js
@@ -1,11 +1,13 @@
 import { MissionUtils } from '@woowacourse/mission-utils';
 
+import { OTHERS } from '../../constants/constants.js';
+
 export const progressRacingCar = async (cars) => {
   cars.forEach((car) => {
-    const carInfo = `${car.name}: ${'-'.repeat(car.go)}`;
+    const carInfo = `${car.name}: ${OTHERS.PROGRESS_BAR.repeat(car.go)}`;
 
     MissionUtils.Console.print(carInfo);
   });
 
-  MissionUtils.Console.print('');
+  MissionUtils.Console.print(OTHERS.NEXT_LINE);
 };
