--- src/view/printWinner.js
@@ -1,5 +1,0 @@
-import { MissionUtils } from '@woowacourse/mission-utils';
-
-export const printWinner = async (winnerNames) => {
-  return MissionUtils.Console.print(`최종 우승자 : ${winnerNames}`);
-};
